requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
opencv-python>=4.10
deepface>=0.0.93
numpy>=1.26
Pillow>=10.3
tensorflow>=2.16
keras>=3.4
python-dotenv>=1.0.1
